+++ src/bos/server/utils.py
@@ -29,7 +29,7 @@ class ParsingException(Exception):
     pass
 
 
-def _canonize_xname(xname):
+def canonize_xname(xname):
     """Ensure the xname is canonical.
     * Its components should be lowercase.
     * Any leading zeros should be stripped off.
@@ -43,7 +43,7 @@ def _canonize_xname(xname):
     return re.sub(r'x0*(\d+)c0*(\d+)s0*(\d+)b0*(\d+)n0*(\d+)', r'x\1c\2s\3b\4n\5', xname.lower())
 
 
-def _get_request_json():
+def get_request_json():
     """
     Used by endpoints which are expecting a JSON payload in the request body.
     Returns the JSON payload.